requirements.txt
numpy
matplotlib
tensorflow
tensorflow_probability
tf_keras
pyserial